Tutorial

Ready-to-run scripts for the functionalities introduced here can be found in the examples directory of the repo.

Creating a bacterium

Bacteria are represented by custom types, that must be subtypes of the AbstractAgent type implemented by Agents.jl.

BacteriaBasedModels.AbstractMicrobe — Type

AbstractMicrobe{D} <: AbstractAgent where {D<:Integer}
YourMicrobeType{D} <: AbstractMicrobe{D}

All microbe types in BacteriaBasedModels.jl simulations must be instances of user-defined types that are subtypes of AbstractMicrobe. The parameter D defines the dimensionality of the space in which the microbe type lives (1, 2 and 3 are currently supported).

All microbe types should have the following fields: - id::Int → id of the microbe - pos::NTuple{D,Float64} → position of the microbe - vel::NTuple{D,Float64} → velocity of the microbe - motility → motile pattern of the microbe - turn_rate::Float64 → average reorientation rate of the microbe - rotational_diffusivity::Float64 → rotational diffusion coefficient

By default, BacteriaBasedModels provides a basic Microbe type, that is usually sufficient for the simplest types of simulations.

BacteriaBasedModels.Microbe — Type

Microbe{D} <: AbstractMicrobe{D}

Basic microbe type for simple simulations.

Default parameters:

id::Int → identifier used internally by Agents.jl

pos::NTuple{D,Float64} = ntuple(zero,D) → position

motility = RunTumble() → motile pattern

vel::NTuple{D,Float64} = rand_vel(D) .* rand(motility.speed) → velocity vector

turn_rate::Float64 = 1.0 → frequency of reorientations

state::Float64 → generic variable for a scalar internal state

rotational_diffusivity::Float64 = 0.0 → rotational diffusion coefficient

radius::Float64 = 0.0 → equivalent spherical radius of the microbe

In order to create a Microbe living in a one-dimensional space we can just call

Microbe{1}(id=0)

It is required to pass a value to the id argument (this behavior might change in the future). All the other parameters will be given default values (as described in the type docstring) if not assigned explicitly.

Similarly, for bacteria living in two or three dimensions we can use

Microbe{2}(id=0)
Microbe{3}(id=0)